refactor: WishException 을 service 에서 domain 패키지로 이동
- 도메인(Wish·WishCursor)이 서비스 계층 예외를 참조하던 계층 역전 해소
- HttpMappable·status 는 커스텀 예외에 박는 컨벤션대로 유지
- 다른 도메인 예외(ProductLinkException)와 패키지 위치 일관성 확보
