+++ src/main/kotlin/com/depromeet/team3/tournament/service/TournamentService.kt
@@ -11,8 +11,8 @@ import com.depromeet.team3.tournament.service.dto.AddTournamentItems
 import com.depromeet.team3.tournament.service.dto.CreateTournament
 import com.depromeet.team3.tournament.service.dto.RecordMatch
 import com.depromeet.team3.tournament.service.dto.TournamentInfo
+import com.depromeet.team3.wishlist.domain.WishException
 import com.depromeet.team3.wishlist.repository.WishRepository
-import com.depromeet.team3.wishlist.service.WishException
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import java.util.UUID
